Completing the documentation with automatic status update on Allegro
In the Allegro panel, you can set automatic status change to "Sent" after entering the tracking number for a given order. You must enter the tracking number for a given order manually or edit an existing one and select the "Automatically change the status of all orders to SHIPPED after adding a tracking number" checkbox.